requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.2
django-cors-headers==4.6.0
filelock==3.16.0
fsspec==2024.9.0
huggingface-hub==0.24.7
MarkupSafe==2.1.5
mpmath==1.3.0
mysqlclient==2.2.5
networkx==3.3
numpy==2.1.1
packaging==24.1
pillow==10.4.0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
setuptools==75.0.0
sqlparse==0.5.1
sympy==1.13.2
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
transformers @ git+https://github.com/huggingface/transformers.git@8bd2b1e8c23234cd607ca8d63f53c1edfea27462
typing_extensions==4.12.2
urllib3==2.2.3
